+++ packages/pigeon/platform_tests/alternate_language_test_plugin/android/src/test/java/com/example/alternate_language_test_plugin/NullableReturnsTest.java
@@ -76,7 +76,7 @@ public class NullableReturnsTest {
     boolean[] didCall = {false};
     api.doit(
         null,
-        new NullableReturns.Result<Long>() {
+        new NullableReturns.NullableResult<Long>() {
           public void success(Long result) {
             didCall[0] = true;
             assertNull(result);