+++ packages/pigeon/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Adds error handling on Flutter Api methods.
 * **Breaking Change** Removes `Reply` class from all Java method returns.
+* **Breaking Change** Adds `NullableResult` class for all nullable Java method returns.
 * **Breaking Change** Nests all enum returns in Objective-c to allow for `nil` response on error.
 
 ## 11.0.1